[BIOMAGE-1155] [BIOMAGE-1164] Fix volcano plot issues (#387)
* Get y-axis to show up

What on earth

* Get title to show up

* Fix properly

* Remove trailing space in y-axis title

* Remove insignificant and unknown statuses

Requested by @ogibson-biomage

* Make sliders slightly functional

* Support "auto" (==0) range

* Remove xMax calculation

It was always too small as a maximum, around
the same as "auto" results in anyway and
expensive to calculate. If the new value being
a constant is bad then then revert and use the
original value multiplied by 10 or something.

* Increase maximum y-axis range

So it's bigger than "auto"

* Fix merge conflict

* Remove default value for single-sample basis

This fixes the problem where "All" cannot be
selected. It additionally fixes the validation
problem in #382.

I can't yet work out yet to keep the default and
properly fix these two issues.

* Set diff expression default sample again

* Don't use range for colour

* Fix title size config reference

* Edit y-axis label